Fix implementation of query to get PackCardInformation

+++ src/main/java/com/rakoon/backend/model/view/PackCardDto.java
@@ -1,5 +1,6 @@
 package com.rakoon.backend.model.view;
 
+import jakarta.persistence.Entity;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -25,5 +26,5 @@ public class PackCardDto {
     //Package Data
     private double packPrice;
     private double packDiscountRate;
-    private Integer quantityPacksAvailable;
+    private Long quantityPacksAvailable;
 }
